Moved llvm setup to main. Updated parser and parser tests to reflect this.
